requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1617740417847/work
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==2.2
django-adaptors==0.2.5
django-azure-sql-backend==2.1.1.1
django-cors-headers==3.7.0
django-extensions==3.1.3
django-filter==2.4.0
django-mssql==1.8
django-mssql-azure-backend==1.0.2
django-mssql-backend==2.8.1
django-mssql-backend-azure==2.9.2
django-mysql==3.12.0
django-pymssql==1.7.1
djangorestframework==3.12.0
idna==2.10
msal==1.12.0
mysqlclient @ file:///tmp/build/80754af9/mysqlclient_1609786120714/work
Pillow==8.2.0
psycopg2 @ file:///tmp/build/80754af9/psycopg2_1612298595717/work
pycparser==2.20
PyJWT==2.1.0
pymssql==2.2.1
pyodbc==4.0.30
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
requests==2.25.1
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
urllib3==1.26.5
